change data type

+++ lib/Model/Student/Marks.php
@@ -10,7 +10,7 @@ class Model_Student_Marks extends Model_Table {
 		$this->hasOne('Exam','exam_id');
 		$this->hasOne('Subject','subject_id');
 		$this->hasOne('Session','session_id');
-		$this->addField('marks')->type('int')->display(array('grid'=>'grid/inline'));
+		$this->addField('marks')->type('money')->display(array('grid'=>'grid/inline'));
 		$this->addHook('beforeSave',$this);
 		$this->add('dynamic_model/Controller_AutoCreator');
 	}
